+++ NEWS.md
@@ -5,6 +5,7 @@
 * I renamed the Dockerfiles and I added a separate Dockerfile for the IceSatR package because the docker image didn't work due to the similar naming of the binder image. I also added (in the Dockerfiles) the "awscli" installation which is required for the vignettes. The ".github/workflows/docker_image.yml" file was adjusted as well.
 * I updated the documentation to account for the migration of OpenAltimetry to EARTHDATA (https://openaltimetry.earthdatacloud.nasa.gov/)
 * I updated the *getTracks()* function
+* I fixed a bug in the *vsi_time_specific_orbits_wkt()* function
 
 
 ## IceSat2R 1.0.4